Updated for new libmicrohttpd
+ The webview plugins now supports and *requires* libmicrohttpd >= 0.4.0. API
. changes are minimal but incompatible so we force the upgrade now. New
. packages for the Nao are available in the naorepo feed. Fedora delivers the
. new version for a few days already.

Tim


git-svn-id: https://lechuck.informatik.rwth-aachen.de:2342/svn/fawkes/trunk@1972 fc82b16a-671c-0410-8732-9903a8d2dd0d

+++ src/plugins/webview/plugins_processor.h
@@ -42,7 +42,7 @@ class WebPluginsRequestProcessor : public WebRequestProcessor
 				     const char *method,
 				     const char *version,
 				     const char *upload_data,
-				     unsigned int *upload_data_size,
+				     size_t *upload_data_size,
 				     void **session_data);
 
  private:
