+++ src/plugins/webview/request_dispatcher.cpp
@@ -60,7 +60,7 @@ WebRequestDispatcher::process_request_cb(void *callback_data,
 					 const char *method,
 					 const char *version,
 					 const char *upload_data,
-					 unsigned int *upload_data_size,
+					 size_t *upload_data_size,
 					 void **session_data)
 {
   WebRequestDispatcher *rd = static_cast<WebRequestDispatcher *>(callback_data);
@@ -146,7 +146,7 @@ WebRequestDispatcher::process_request(struct MHD_Connection * connection,
 				      const char *method,
 				      const char *version,
 				      const char *upload_data,
-				      unsigned int *upload_data_size,
+				      size_t *upload_data_size,
 				      void **session_data)
 {
   std::string surl = url;
