+++ src/plugins/webview/request_processor.h
@@ -35,7 +35,7 @@ class WebRequestProcessor
 				     const char *method,
 				     const char *version,
 				     const char *upload_data,
-				     unsigned int *upload_data_size,
+				     size_t *upload_data_size,
 				     void **session_data)               = 0;
 
 };
