+++ src/plugins/webview/startpage_processor.cpp
@@ -58,7 +58,7 @@ WebStartPageRequestProcessor::process_request(const char *url,
 					       const char *method,
 					       const char *version,
 					       const char *upload_data,
-					       unsigned int *upload_data_size,
+					       size_t *upload_data_size,
 					       void **session_data)
 {
   if ( strncmp("/", url, 1) == 0 ) {
