+++ src/plugins/webview/startpage_processor.h
@@ -41,7 +41,7 @@ class WebStartPageRequestProcessor : public WebRequestProcessor
 				     const char *method,
 				     const char *version,
 				     const char *upload_data,
-				     unsigned int *upload_data_size,
+				     size_t *upload_data_size,
 				     void **session_data);
 
  private:
