+++ src/plugins/webview/static_processor.h
@@ -44,7 +44,7 @@ class WebStaticRequestProcessor : public WebRequestProcessor
 				     const char *method,
 				     const char *version,
 				     const char *upload_data,
-				     unsigned int *upload_data_size,
+				     size_t *upload_data_size,
 				     void **session_data);
 
  private:
